move Content to components
